Separate out section logic from other regtext
Also rewrites associated tests in py.test style
