+++ regparser/commands/current_version.py
@@ -7,7 +7,7 @@ from regparser.history.annual import find_volume
 from regparser.history.versions import Version
 from regparser.index import dependency, entry
 from regparser.notice.fake import build as build_fake_notice
-from regparser.tree import xml_parser
+from regparser.tree import gpo_cfr
 
 
 _version_id = '{}-annual-{}'.format
@@ -26,7 +26,7 @@ def process_if_needed(volume, cfr_part):
     deps.add(tree_entry, annual_entry)
     deps.validate_for(tree_entry)
     if deps.is_stale(tree_entry):
-        tree = xml_parser.reg_text.build_tree(annual_entry.read().xml)
+        tree = gpo_cfr.builder.build_tree(annual_entry.read().xml)
         tree_entry.write(tree)
         notice_entry.write(build_fake_notice(
             version_id, volume.publication_date, volume.title, cfr_part))
