+++ regparser/tree/gpo_cfr/appendices.py
@@ -1,3 +1,4 @@
+# -*- coding: utf-8 -*-
 from __future__ import unicode_literals
 from itertools import takewhile
 import logging
